Stop cargo t from testing README code that isn't rust, remove make_logger entirely from the serialkeel python package
