+++ py/examples/test_mock_concurrent.py
@@ -1,10 +1,31 @@
 import pytest
 from pathlib import Path
 
-from serialkeel import connect, make_logger
+from serialkeel import connect
 import os
+import logging
 
 # See README.md for information on how to run this.
+
+def make_logger(name: str, add_formatter: bool = False) -> logging.Logger:
+    logger = logging.getLogger(f'{name}')
+
+    logdir = Path('logs')
+    logdir.mkdir(parents=True, exist_ok=True)
+
+    logfile = logdir / f'{name}.log'
+
+    h = logging.FileHandler(logfile, mode='w')
+    if add_formatter:
+        h.setFormatter(logging.Formatter('%(asctime)s [%(levelname)s] %(message)s'))
+    h.setLevel(logging.DEBUG)
+    logger.setLevel(logging.DEBUG)
+
+    # Note that depending on how many tests parameterized,
+    # this almost doubles time executed
+    logger.addHandler(h)  # <--
+
+    return logger
 
 @pytest.mark.asyncio_cooperative
 @pytest.mark.parametrize("n", range(10))
